pyproject.toml
[tool.poetry]
name = "bintweet"
version = "0.1.0"
description = ""
authors = ["tulis"]

[tool.poetry.dependencies]
python = "^3.8"
tweepy = "^3.10.0"
fire = "^0.4.0"
loguru = "^0.5.3"
python-dotenv = "^0.15.0"
iso8601 = "^0.1.14"

[tool.poetry.dev-dependencies]
pytest = "^4.6"
black = "^20.8b1"
isort = "^5.7.0"
flake8 = "^3.8.4"
mypy = "^0.800"
pre-commit = "^2.10.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 80
